pyproject.toml
[project]
name = "pythonexercise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jieba>=0.42.1",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "wordcloud>=1.9.4",
]
